% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandAltmanWrapper.R
\name{blandAltman}
\alias{blandAltman}
\title{Bland-Altman Plots}
\usage{
blandAltman(
  data = NULL,
  version = "1",
  blandAltmanTable = FALSE,
  ci = FALSE,
  ciLevel = 0.95,
  ciShading = FALSE,
  ciShadingWithColour = FALSE,
  pairs = list(),
  plotHeight = 320,
  plotWidth = 480
)
}
\arguments{
\item{blandAltmanTable, }{Table showing the mean difference, the upper and lower limits of agreement, and their confidence intervals.
Defaults to \code{FALSE}.}

\item{ci, }{Display confidence intervals around the mean difference and its upper and lower limits of agreement in the plot.
Defaults to \code{FALSE}.}

\item{ciShading, }{Shade the confidence interval regions around the mean difference and limits of agreement.
Defaults to \code{FALSE}.}

\item{ciShadingWithColour, }{Use colour (rather than grey) to fill the shaded confidence regions.
Defaults to \code{FALSE}.}

\item{pairs, }{Pairs of scale variables to compare. Each pair produces a Bland-Altman plot of the mean versus the difference between measurements.}
}
\description{
Bland-Altman Plots
}
