% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidimensionalReliabilityFrequentist.R
\name{.frequentistOmegaScale}
\alias{.frequentistOmegaScale}
\title{So the structure here is to first calculcate the bootstrap samples per coefficient
And in the computeScaleResults function to summarize the results}
\usage{
.frequentistOmegaScale(jaspResults, dataset, options, model)
}
\description{
So the structure here is to first calculcate the bootstrap samples per coefficient
And in the computeScaleResults function to summarize the results
}
