% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardErrorOfMeasurement.R
\name{.semComputeCis}
\alias{.semComputeCis}
\title{this function computes the confidence intervals for the sum scores
computes two CIs, one for the table and one for the plot in JASP
out is the output matrix from the sem functions, that is, it contains the sem values
scores is a vector with the sum scores
ciLevelTable is the confidence level for the table
ciLevelPlots is the confidence level for the plots}
\usage{
.semComputeCis(out, scores, ciLevelTable, ciLevelPlots)
}
\description{
this function computes the confidence intervals for the sum scores
computes two CIs, one for the table and one for the plot in JASP
out is the output matrix from the sem functions, that is, it contains the sem values
scores is a vector with the sum scores
ciLevelTable is the confidence level for the table
ciLevelPlots is the confidence level for the plots
}
