% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardErrorOfMeasurement.R
\name{.semComputeWithCaseMin}
\alias{.semComputeWithCaseMin}
\title{this function computes the SEMs for the different score groups
there is a special treatment to collapse some scores into groups if they do not reach the min size
out is the output matrix from the previous function
S is a vector with the sum scores
partSUMS is a matrix with the sum scores of the test parts
fun is the function that computes the SEM, this one differs for each method
cc is a vector with the number of items in each test part, this is only needed for the Lord's compound method}
\usage{
.semComputeWithCaseMin(out, S, caseMin, partSUMS, fun, cc = NULL)
}
\description{
this function computes the SEMs for the different score groups
there is a special treatment to collapse some scores into groups if they do not reach the min size
out is the output matrix from the previous function
S is a vector with the sum scores
partSUMS is a matrix with the sum scores of the test parts
fun is the function that computes the SEM, this one differs for each method
cc is a vector with the number of items in each test part, this is only needed for the Lord's compound method
}
