% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardErrorOfMeasurement.R
\name{.semCounts}
\alias{.semCounts}
\title{X is the data matrix
nc is the number of response categories
K is the number of splits
this function counts the sum scores and returns the scores and their counts}
\usage{
.semCounts(X)
}
\description{
X is the data matrix
nc is the number of response categories
K is the number of splits
this function counts the sum scores and returns the scores and their counts
}
