% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardErrorOfMeasurement.R
\name{.semOptionsHelper}
\alias{.semOptionsHelper}
\title{this might be the most important function :-}
\usage{
.semOptionsHelper(options, dataset, ready)
}
\description{
this might be the most important function :-
}
