% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardErrorOfMeasurement.R
\name{.semPrepareOutMatrix}
\alias{.semPrepareOutMatrix}
\title{this function prepares the output matrix for the SEM computations}
\usage{
.semPrepareOutMatrix(nit, nc, scrs)
}
\description{
this function prepares the output matrix for the SEM computations
}
