% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardErrorOfMeasurement.R
\name{.semPrepareSumScores}
\alias{.semPrepareSumScores}
\title{this function prepares the sum scores for the SEM computations:
for some functions it implements the split and saves the sum scores of the test parts}
\usage{
.semPrepareSumScores(X, K)
}
\description{
this function prepares the sum scores for the SEM computations:
for some functions it implements the split and saves the sum scores of the test parts
}
