% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardErrorOfMeasurement.R
\name{.semThorn}
\alias{.semThorn}
\title{if not stated otherwise:
X is the data matrix
K is the number of splits
nc is the number of response categories
caseMin is the minimum number of cases in a group
scrs is a vector of sum scores}
\usage{
.semThorn(X, nc, caseMin, scoresObj)
}
\description{
if not stated otherwise:
X is the data matrix
K is the number of splits
nc is the number of response categories
caseMin is the minimum number of cases in a group
scrs is a vector of sum scores
}
